Move decorateWithCollation to utils, move test to collations_tests
