+++ lib/collection.js
@@ -13,6 +13,7 @@ const toError = require('./utils').toError;
 const normalizeHintField = require('./utils').normalizeHintField;
 const handleCallback = require('./utils').handleCallback;
 const decorateCommand = require('./utils').decorateCommand;
+const decorateWithCollation = require('./utils').decorateWithCollation;
 const formattedOrderClause = require('./utils').formattedOrderClause;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const CommandCursor = require('./command_cursor');
@@ -2278,17 +2279,6 @@ var findAndRemove = function(self, query, sort, options, callback) {
   // Execute the callback
   self.findAndModify(query, sort, null, options, callback);
 };
-
-function decorateWithCollation(command, self, options) {
-  // Do we support collation 3.4 and higher
-  var capabilities = self.s.topology.capabilities();
-  // Do we support write concerns 3.4 and higher
-  if (capabilities && capabilities.commandsTakeCollation) {
-    if (options.collation && typeof options.collation === 'object') {
-      command.collation = options.collation;
-    }
-  }
-}
 
 function decorateWithReadConcern(command, self, options) {
   let readConcern = Object.assign({}, command.readConcern || {});
