+++ lib/cursor.js
@@ -5,6 +5,7 @@ const f = require('util').format;
 const deprecate = require('util').deprecate;
 const formattedOrderClause = require('./utils').formattedOrderClause;
 const handleCallback = require('./utils').handleCallback;
+const decorateWithCollation = require('./utils').decorateWithCollation;
 const ReadPreference = require('mongodb-core').ReadPreference;
 const MongoError = require('mongodb-core').MongoError;
 const Readable = require('stream').Readable;
@@ -1010,9 +1011,7 @@ var count = function(self, applySkipLimit, opts, callback) {
   }
 
   // Apply a collation if set
-  if (self.s.cmd.collation) {
-    command.collation = self.s.cmd.collation;
-  }
+  decorateWithCollation(command, self, self.s.cmd);
 
   if (typeof opts.maxTimeMS === 'number') {
     command.maxTimeMS = opts.maxTimeMS;
