+++ lib/utils.js
@@ -505,6 +505,26 @@ function isPromiseLike(maybePromise) {
   return maybePromise && typeof maybePromise.then === 'function';
 }
 
+/**
+ * Applies collation to a given command.
+ *
+ * @param {object} [command] the command on which to apply collation
+ * @param {(Cursor|Collection)} [self] 
+ * @param {object} [options] options containing collation settings
+ */
+function decorateWithCollation(command, self, options) {
+  if (self.s && self.s.topology) {
+    // Do we support collation 3.4 and higher
+    var capabilities = self.s.topology.capabilities();
+    // Do we support write concerns 3.4 and higher
+    if (capabilities && capabilities.commandsTakeCollation) {
+      if (options.collation && typeof options.collation === 'object') {
+        command.collation = options.collation;
+      }
+    }
+  }
+}
+
 exports.filterOptions = filterOptions;
 exports.mergeOptions = mergeOptions;
 exports.translateOptions = translateOptions;
@@ -525,3 +545,4 @@ exports.translateReadPreference = translateReadPreference;
 exports.executeOperation = executeOperation;
 exports.applyWriteConcern = applyWriteConcern;
 exports.isPromiseLike = isPromiseLike;
+exports.decorateWithCollation = decorateWithCollation;
